ランタイムコピー用バッチファイル、ビルドバッチファイルの追加、README更新
* README にある「環境構築とSLANGプログラムのビルド方法」に、ツール類のインストール、ディスクイメージの準備、ビルドバッチの実行方法などを記載しました
