+++ copyruntime.bat
@@ -1,0 +1,5 @@
+ECHO OFF
+
+mkdir %homedrive%%homepath%\.config\SLANG\
+copy *.env %homedrive%%homepath%\.config\SLANG\
+copy *.yml %homedrive%%homepath%\.config\SLANG\
