+++ slbuild.bat
@@ -1,0 +1,120 @@
+ECHO OFF
+
+IF "%1" EQU "" GOTO REQUIREFILE
+
+REM program info
+SET CURPATH=%~dp0
+SET PROGPATH=%1
+CALL :set_progpath %PROGPATH%
+
+REM Target = lsx / x1 / sos / msx2
+IF "%2" EQU "" (
+SET TARGETENV=lsx
+SET CHECKCPM=0
+) ELSE IF "%2" EQU "cpm" (
+SET TARGETENV=lsx
+SET CHECKCPM=1
+) ELSE (
+SET TARGETENV=%2
+SET CHECKCPM=0
+)
+
+REM Additional library
+REM SET ADDLIB=-L soroban
+
+REM use cpm.exe (env: lsx)
+
+REM runtime copy to ~/.config/SLANG
+SET COPYRUNTIME=0
+SET RUNTIMEPATH=.
+
+REM S-OS addresses
+SET LOADADR=3000
+SET RUNADR=3000
+
+SET TOOLPATH=%CURPATH%tools
+SET IMAGEPATH=%CURPATH%images
+
+SET SLANGCOMPILER=%CURPATH%bin\SLANGCompiler.exe
+SET ASM=%TOOLPATH%\AILZ80ASM.exe
+
+SET EMULATOR=C:\emu\x1\x1.exe
+
+SET NDCPATH=%TOOLPATH%\NDC.exe
+SET HUDISKPATH=%TOOLPATH%\HuDisk.exe
+SET CPMEMUPATH=%TOOLPATH%\cpm.exe
+
+IF %COPYRUNTIME% gtr 0 (
+	mkdir %homedrive%%homepath%\.config\SLANG\
+	copy %RUNTIMEPATH%\*.env %homedrive%%homepath%\.config\SLANG\
+	copy %RUNTIMEPATH%\*.yml %homedrive%%homepath%\.config\SLANG\
+)
+
+cd %PROGDIR%
+
+%SLANGCOMPILER% %PROG%%PROGEXT% -E %TARGETENV% %ADDLIB%
+IF not %ERRORLEVEL% == 0 GOTO ERROR
+
+%ASM% %PROG%.ASM -sym -lst -bin -f
+IF not %ERRORLEVEL% == 0 GOTO ERROR
+
+IF %TARGETENV%==lsx GOTO LSXENV
+IF %TARGETENV%==x1 GOTO LSXENV
+IF %TARGETENV%==sos GOTO SOSENV
+IF %TARGETENV%==msx2 GOTO NOTSUPPORTED
+
+:SOSENV
+ECHO S-OS
+SET PROGIMAGE=SOSPROG.D88
+DEL PROG.BIN
+REN %PROG%.BIN PROG.BIN
+%HUDISKPATH% -d %IMAGEPATH%\%PROGIMAGE% PROG.BIN
+%HUDISKPATH% -a %IMAGEPATH%\%PROGIMAGE% PROG.BIN -r %LOADADR% -g %RUNADR%
+
+GOTO DONE
+:LSXENV
+ECHO LSX-Dodgers
+
+IF %CHECKCPM% gtr 0 GOTO CHECKCPMEMU
+
+SET PROGIMAGE=LSXPROG.D88
+
+DEL PROG.COM
+REN %PROG%.BIN PROG.COM
+%NDCPATH% D %IMAGEPATH%\%PROGIMAGE% 0 PROG.COM
+%NDCPATH% P %IMAGEPATH%\%PROGIMAGE% 0 PROG.COM
+
+:DONE
+cd %CURPATH%
+
+%EMULATOR% %IMAGEPATH%\%PROGIMAGE%
+
+ECHO DONE!
+EXIT /B
+
+:NOTSUPPORTED
+ECHO NOT SUPPORTED %TARGETENV%
+
+:ERROR
+ECHO ERROR!
+cd %CURPATH%
+
+:CHECKCPMEMU
+%CPMEMUPATH% %PROG%.BIN
+cd %CURPATH%
+EXIT /B
+
+
+:REQUIREFILE
+ECHO SLANG Compile batch v.1.0
+ECHO  slbuild.bat SLANGSource.SL
+EXIT /B
+
+:set_progpath
+
+SET PROGDIR=%~dp1
+SET PROG=%~n1
+SET PROGEXT=%~x1
+
+ECHO SOURCE : %PROGDIR%%PROG%%PROGEXT%
+
